+++ src/Schema/Factories/DirectiveFactory.php
@@ -61,18 +61,26 @@ class DirectiveFactory
     public function __construct(Dispatcher $dispatcher)
     {
         // The namespaces will be tried in the order that they contain
-        $this->directiveBaseNamespaces = (new Collection([
+        $directives = new Collection([
             // User defined directives (top priority)
             config('lighthouse.namespaces.directives'),
 
             // Plugin developers defined directives
             $dispatcher->dispatch(new RegisterDirectiveNamespaces),
+        ]);
 
-            // Lighthouse defined directives
-            'Nuwave\\Lighthouse\\Schema\\Directives',
-        ]))->flatten()
-           ->filter()
-           ->all();
+        // Allow a smooth transition away from the deprecated directives
+        if(config('new_between_directives')){
+            $directives->push('Nuwave\\Lighthouse\\Between');
+        }
+
+        // Lighthouse defined directives
+        $directives->push('Nuwave\\Lighthouse\\Schema\\Directives');
+
+        $this->directiveBaseNamespaces = $directives
+            ->flatten()
+            ->filter()
+            ->all();
     }
 
     /**