+++ tests/Integration/Execution/BuilderTest.php
@@ -1,12 +1,19 @@
 <?php
 
-namespace Tests\Integration\Schema\Directives;
+namespace Tests\Integration\Execution;
 
 use Tests\DBTestCase;
 use Tests\Utils\Models\User;
 
-class QueryFilterDirectiveTest extends DBTestCase
+class BuilderTest extends DBTestCase
 {
+    protected function getEnvironmentSetUp($app)
+    {
+        parent::getEnvironmentSetUp($app);
+
+        $app['config']->set('new_between_directives', true);
+    }
+
     /**
      * @var \Illuminate\Database\Eloquent\Collection<\Tests\Utils\Models\User>
      */